+++ packages/just-scripts/src/tasks/index.ts
@@ -14,4 +14,5 @@ export * from './prettierTask';
 export * from './eslintTask';
 export * from './webpackCliInitTask';
 export * from './webpackCliTask';
+export * from './webpackDevServerTask';
 export * from './tarTask';
